Adding function to draw lines.

+++ types/line.h
@@ -6,6 +6,6 @@
 namespace Px {
     struct Line {
         public:
-            uint32_t x1, y1, x2, y2;
+            float x1, y1, x2, y2;
     };
 } // namespace Px
